requirements.txt
aiofiles==0.5.0
face-alignment==1.0.0
fastapi==0.54.2
imageio==2.8.0
opencv-python==4.2.0.34
pandas==1.0.3
Pillow==7.1.2
python-multipart==0.0.5
PyYAML==5.3.1
requests==2.23.0
scikit-image==0.17.2
scikit-learn==0.23.0
torch==1.1.0
torchvision==0.3.0
tqdm==4.46.0
uvicorn==0.11.5
uvloop==0.14.0
websockets==8.1
tensorflow-cpu
python_path
imageio-ffmpeg